changed datatype of quantity attribute (Integer to BigDecimal)
git-svn-id: http://emr.moh.gov.rw/svn/modules/moh-billing/branches/1.6.x@3195 6101c9c6-cf70-4d46-9f71-e9b75042c301

+++ src/org/openmrs/module/mohbilling/model/Consommation.java
@@ -1,12 +1,13 @@
 package org.openmrs.module.mohbilling.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Consommation {
 	Date recordDate;
 	String libelle;
 	Double unitCost;
-	Integer quantity;
+	BigDecimal quantity;
 	Double cost;
 	Double patientCost;
 	Double insuranceCost;
@@ -37,13 +38,13 @@ public class Consommation {
 	/**
 	 * @return the quantity
 	 */
-	public Integer getQuantity() {
+	public BigDecimal getQuantity() {
 		return quantity;
 	}
 	/**
 	 * @param quantity the quantity to set
 	 */
-	public void setQuantity(Integer quantity) {
+	public void setQuantity(BigDecimal quantity) {
 		this.quantity = quantity;
 	}
 	/**
